# src/main/resources/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Parliament Browser API",
    "version": "1.0.0",
    "description": "Eine API zur Interaktion mit dem Parliament Browser Projekt."
  },
  "host": "localhost:4567",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
...
    "/reden/:redeId": {
      "get": {
        "summary": "Ruft eine spezifische Rede ab",
        "description": "Ruft Details zu einer spezifischen Rede basierend auf ihrer ID ab.",
        "parameters": [
          {
            "name": "redeId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Die ID der abzurufenden Rede."
          }
        ],
        "responses": {
          "200": {
            "description": "Details der spezifischen Rede in einer gerenderten HTML-Seite.",
            "schema": {
              "type": "string"
            }
          },
          "404": {
            "description": "Rede nicht gefunden."
          }
        }
      }
    },
    "/rede/text": {
      "get": {
        "summary": "Sucht Reden basierend auf Text",
        "description": "Sucht nach Reden, die einen bestimmten Text enthalten.",
        "parameters": [
          {
            "name": "textQuery",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "Der Text, nach dem innerhalb der Reden gesucht werden soll."
          }
        ],
        "responses": {
          "200": {
            "description": "Eine Liste von Reden, die den Text enthalten, dargestellt auf einer HTML-Seite.",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
    "/redenSearch/:sitzungsNr/:top": {
      "get": {
        "summary": "Suche nach Reden basierend auf TOP",
        "description": "Stellt eine Suche nach Reden basierend auf einem spezifischen Tagesordnungspunkt (TOP) bereit.",
        "parameters": [
          {
            "name": "sitzungsNr",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Die Sitzungsnummer, zu der der TOP gehört."
          },
          {
            "name": "top",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Der Tagesordnungspunkt, nach dem gesucht werden soll."
          }
        ],
        "responses": {
          "200": {
            "description": "Ergebnisse der Suche, dargestellt auf einer HTML-Seite.",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
  }
}